Indiv-level output, variation in tgt-neighbors
primates within a simulation can now have different tgt-neighbor values that they follow. this is controlled by a switch. there is also a way to get individual data (also controlled by a switch) so that the differing tgt-neighbor values can be assessed for their impact on the individual's foraging success.

updated to run on latest version of NetLogo