adds extensions, multiple tgt-neighbor-setting
adds the extensions to the model folder in attempt to enable logolink (package for running netlogo from R). also expands the variation in tgt-neighbor to include 4 different permutations. also, now priamtes only cohere with those with the same tgt-neighbor value as them, to try to create small groups and large groups on the same landscape.